diffMockModel
